tests: add yarn script
