+++ package.json
@@ -20,7 +20,9 @@
     "gen:func-js": "ts-node src/func/func.build.ts",
     "gen:contracts:examples": "ts-node examples/contracts.build.ts",
     "gen:contracts:test": "ts-node src/test/contracts.build.ts",
-    "gen:contracts:benchmarks": "ts-node src/test/benchmarks/benchmarks.build.ts",
+    "gen:contracts:benchmarks:tact": "ts-node src/test/benchmarks/benchmarks.build.ts",
+    "gen:contracts:benchmarks:func": "ts-node src/test/benchmarks/func.build.ts",
+    "gen:contracts:benchmarks": "yarn gen:contracts:benchmarks:tact && yarn gen:contracts:benchmarks:func",
     "gen:contracts:test:map": "ts-node ./src/test/e2e-emulated/map-tests/generate.ts",
     "gen:contracts:all": "yarn gen:contracts:examples && yarn gen:contracts:test && yarn gen:contracts:benchmarks && yarn gen:contracts:test:map",
     "gen": "yarn gen:grammar && yarn gen:stdlib && yarn gen:func-js && yarn gen:contracts:all",
