+++ spell/cspell-list.txt
@@ -174,3 +174,5 @@ xffff
 XMODEM
 xtwitter
 привет
+extracurrency
+nobounce
